Use Bull queue for syncing trace campaigns and projects
