+++ config/default.json
@@ -12,6 +12,10 @@
     "host": "localhost",
     "port": 6379
   },
+  "sharedRedis": {
+    "host": "localhost",
+    "port": 6379
+  },
   "bridgeMonitorBaseUrl": "https://feathers.bridge.beta.giveth.io",
   "nedb": "../data",
   "uploads": "../uploads",
@@ -195,11 +199,5 @@
   },
   "givethIoUrl": "http://localhost:4000/graphql",
   "givethIoUpdateProjectUrl": "http://localhost:4000/traceableProject",
-  "givethIoProjectsReviewerAddress": "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1",
-  "givethIoInfo": {
-    "username": "username",
-    "password": "password",
-    "givethIoUsername": "traceUserName",
-    "givethIoPassword": "tracePassword"
-  }
+  "givethIoProjectsReviewerAddress": "0x90F8bf6A479f320ead074411a4B0e7944Ea8c9C1"
 }
