+++ src/adapters/givethIo/givethIoMockAdapter.js
@@ -31,8 +31,5 @@ const getUserByUserId = _userId => {
     location: 'world',
   };
 };
-// eslint-disable-next-line no-unused-vars
-const updateGivethIoProject = ({ title, description, campaignId, givethIoProjectId }) => {
-  return true;
-};
-module.exports = { getProjectInfoBySLug, getUserByUserId, updateGivethIoProject };
+
+module.exports = { getProjectInfoBySLug, getUserByUserId };
