--- src/services/verifiedCampaigns/verifiedCampaigns.hooks.js
@@ -1,7 +1,0 @@
-const { authorizeGivethio } = require('../../hooks/authorizeGivethio');
-
-module.exports = {
-  before: {
-    update: [authorizeGivethio()],
-  },
-};
