+++ src/services/verifiedCampaigns/verifiedCampaigns.service.js
@@ -1,12 +1,13 @@
 const errors = require('@feathersjs/errors');
 const config = require('config');
 const logger = require('winston');
-const { NotFound } = require('@feathersjs/errors');
 const { findCampaignByGivethIoProjectId } = require('../../repositories/campaignRepository');
 const { getGivethIoAdapter } = require('../../adapters/adapterFactory');
-const { convertGivethIoToTraceImage } = require('../../utils/givethIoUtils');
-const { CampaignStatus } = require('../../models/campaigns.model');
-const hooks = require('./verifiedCampaigns.hooks');
+
+const convertGivethIoToTraceImage = image => {
+  const imageIpfsPath = image.match(/\/ipfs\/.*/);
+  return imageIpfsPath ? imageIpfsPath[0] : image;
+};
 
 const givethIoAdapter = getGivethIoAdapter();
 module.exports = function verifiedCampaigns() {
@@ -56,37 +57,6 @@ module.exports = function verifiedCampaigns() {
       }
       return { ...projectInfo, owner, reviewerAddress: config.givethIoProjectsReviewerAddress };
     },
-
-    async update(data, params) {
-      const { campaignId, verified, archived, title, description, image } = params;
-      const campaign = await app.service('campaigns').get(campaignId);
-      if (!campaign) {
-        throw new NotFound();
-      }
-      const updateData = {};
-      if (verified !== undefined) {
-        updateData.verified = Boolean(verified);
-      }
-      if (title) {
-        updateData.title = title;
-      }
-      if (description) {
-        updateData.description = description;
-      }
-      if (archived === true && campaign.status === CampaignStatus.ACTIVE) {
-        updateData.status = CampaignStatus.ARCHIVED;
-      } else if (archived === false && campaign.status === CampaignStatus.ARCHIVED) {
-        updateData.status = CampaignStatus.ACTIVE;
-      }
-      if (image) {
-        updateData.image = convertGivethIoToTraceImage(image);
-      }
-      logger.info('update campaign ', updateData);
-      const result = await app.service('campaigns').patch(campaign._id, updateData, {
-        calledFromGivethIo: true,
-      });
-      return result;
-    },
   };
 
   service.docs = {
@@ -134,5 +104,4 @@ module.exports = function verifiedCampaigns() {
     },
   };
   app.use('/verifiedCampaigns', service);
-  app.service('verifiedCampaigns').hooks(hooks);
 };
