--- src/utils/givethIoUtils.js
@@ -1,8 +1,0 @@
-const convertGivethIoToTraceImage = image => {
-  const imageIpfsPath = image.match(/\/ipfs\/.*/);
-  return imageIpfsPath ? imageIpfsPath[0] : image;
-};
-
-module.exports = {
-  convertGivethIoToTraceImage,
-};
